pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stt-quiz-service"
version = "0.1.0"
description = "STT-based quiz and study guide generation service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.135.1",
  "python-dotenv>=1.1.1",
  "requests>=2.32.5",
  "streamlit>=1.55.0",
  "sqlalchemy>=2.0.48",
  "psycopg[binary]>=3.3.3",
  "pgvector>=0.4.2",
  "pydantic>=2.12.5",
  "pydantic-ai>=1.67.0",
  "scikit-learn>=1.7.2",
  "sentence-transformers>=5.1.2",
  "transformers>=4.57.1",
  "uvicorn>=0.41.0",
  "keybert>=0.9.0",
  "yake>=0.6.0",
  "langchain>=0.3.27",
  "langchain-core>=0.3.76",
  "langchain-openai>=0.3.33",
  "langchain-postgres>=0.0.15",
  "langchain-text-splitters>=0.3.11",
  "langchain-google-genai>=2.1.10",
  "langsmith>=0.4.20",
  "tiktoken>=0.11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "httpx>=0.28.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
